buildin/cv/classification/crnn_pytorch/requirement.txt
lmdb==0.97
numpy==1.17.2
Pillow==6.1.0
six==1.12.0





